FilterableProductTable file included
